// petforest-fe-cms/src/utils/enums/home/home.enums.ts
export enum SELECT_COLLAPSE_ENUMS {
  COLLAPSE = 'collapse',
  EXPAND = 'expand',
}

export enum ANNOUNCEMENT {
  FUNERAL = 'FUNERAL',
  NOT_ACCOMPANY = 'NOT_ACCOMPANY',
  ETC = 'ETC',
}

export enum ANNOUNCEMENT_INACTIVE {
  FUNERAL = 'FUNERAL_INACTIVE',
  NOT_ACCOMPANY = 'NOT_ACCOMPANY_INACTIVE',
  ETC = 'ETC_INACTIVE',
}